Из нижней стенки вылетает один шарик, который в некоторый момент взрывается на много шариков

+++ SalutWinFormsApp/SalutBall.cs
@@ -5,11 +5,12 @@ namespace SalutWinFormsApp
     public class SalutBall : MoveBall
     {
         private float g = 0.2f;
-        public SalutBall(Form form, int centerX, int centerY) : base(form)
+        public SalutBall(Form form, float centerX, float centerY) : base(form)
         {
             radius = 15;
             this.centerX = centerX;
             this.centerY = centerY;
+
             vy = -Math.Abs(vy);
         }
